move a bunch of functions to files
